+++ der_floh-ore_value_mod/extensions/scripts/Gvars.gd
@@ -18,7 +18,8 @@ func reset_resources() -> void:
 			_sellable.resource_path, "", ResourceLoader.CacheMode.CACHE_MODE_REPLACE
 		)
 		item.itemSellPrice *= multiplier
-		ModLoaderLog.info(
-			"%s sell price x%.0f = %.0f" % [_sellable.itemID, multiplier, item.itemSellPrice],
-			LOG_NAME
-		)
+		if DerFlohOreValueMod._debug:
+			ModLoaderLog.info(
+				"%s sell price x%.0f = %.0f" % [_sellable.itemID, multiplier, item.itemSellPrice],
+				LOG_NAME
+			)